+++ parser/expansor.c
@@ -6,7 +6,7 @@
 /*   By: antofern <antofern@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 15:45:35 by antofern          #+#    #+#             */
-/*   Updated: 2025/01/15 23:31:01 by antofern         ###   ########.fr       */
+/*   Updated: 2025/01/16 13:23:53 by antofern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,116 +29,9 @@ int	add_pathname(char **cmd,  t_garbage *garbage, char *envp[])
 */
 
 //puede que innecesaria
-int	is_expansible(char *str)
-{
-	if(!str)
-		return (0);
-	while (*str)
-	{
-		if (*str == '$')
-			return (1);
-		if (*str == 39 && ft_strchr(str + 1, 39))
-			str = ft_strchr(str + 1, 39);
-		str++;
-	}
-	return (0);
-}
-
-int	count_expansions(t_task *node)
-{
-	int	count;
-	int	i;
-	
-	count = 1;//reservamos al menos uno extra para add_pathname(), aunque pueda quedar sin uso en algunos casos.
-	count += is_expansible(node->cmd);
-	count += is_expansible(node->redir.infoo);
-	count += is_expansible(node->redir.outfile);
-
-	i = 0;
-	while ((node->argv)[i])
-	{
-		count += is_expansible((node->argv[i]));
-		i++;
-	}
-	return (count);
-}
 
 
-static int	calculate_variable_length(char **str, char *envp[])
-{
-	int j = 0;
-	char *aux;
-	int len;
-
-	(*str)++;
-	while ((*str)[j] && !ft_strchr(WHITESPACES, (*str)[j])
-	&& (*str)[j] != '"' && (*str)[j] != 39 && (*str)[j] != '$') //echo $VAR$VAR es valido.
-		j++;
-	aux = ft_substr(*str, 0, j);
-	if (aux == NULL)
-		return (-1);
-	*str += j;
-	len = ft_strlen(ft_getenv(aux, envp));
-	if (!ft_strcmp(aux,""))
-		len++;
-	free(aux);
-	return len;
-}
-
-//NO SEPARAR DE EXPANDSTR()
-static int	calculate_expansion_length(char *str, char *envp[])
-{
-	int len;
-	int	ret;
-	char *aux;
-
-	len = 0;
-	while (*str)
-	{
-		if (*str == 39 && ft_strchr(str + 1, 39))
-		{
-			aux = str;
-			str = ft_strchr(str + 1, 39);
-			len += (str - aux);
-		}
-		else if(*str == '$')
-		{
-			ret = calculate_variable_length(&str, envp);//consume en str $NOMBRE_DE_VARIABLE
-			if (ret < 0)
-				return (-1);
-			len += ret;
-			continue ;
-		}
-		str++;
-		len++;
-	}
-	return (len);
-}
-
-int handle_dollar(char **new_str, char **str, char **marker, char *envp[])
-{
-	char *aux;
-
-	ft_strlcpy(*new_str, *str, *marker - *str + 1);
-	*new_str = *new_str + (*marker - *str);
-	(*marker)++;
-	*str = *marker;
-	while (**marker && !ft_strchr(WHITESPACES, **marker)
-	&& **marker != '"' && **marker != 39 && **marker != '$')
-		(*marker)++;
-	aux = ft_substr(*str, 0, *marker - *str);
-	if (aux == NULL)
-		return (1);
-	*str = *marker;
-	ft_strlcpy(*new_str, ft_getenv(aux, envp), ft_strlen(ft_getenv(aux, envp)) + 1);
-	if (!ft_strcmp(aux, ""))
-		**new_str='$';
-	free(aux);
-	*new_str = *new_str + ft_strlen(*new_str);
-	return (0);
-}
-
-int	prepare_expansion(char **origin, char **new_str, t_garbage *garbage, char *envp[])
+static int	prepare_expansion(char **origin, char **new_str, t_garbage *garbage, char *envp[])
 {
 	int	len;
 
@@ -189,7 +82,7 @@ int	expandstr(char **origin, t_garbage *garbage, char *envp[]) //envp debe recib
     return 0;
 }
 
-int	expand_task(t_task *node, char *envp[])
+static int	expand_task(t_task *node, char *envp[])
 {
 	int	i;
