+++ parser/str_utils.c
