+++ test/test-expandstr.c
